+++ changelog.md
@@ -1,2 +1,6 @@
+# 1.0.1
+* Removed leaderboard button
+* Fixed created levels info button bug
+
 # 1.0.0
 * Release